requirements.txt
Django==2.2
flake8==3.8.3
flake8-html==0.4.1
pylint==2.5.3
yapf==0.30.0
requests==2.24.0
responses==0.10.15
gunicorn==20.0.4
mysqlclient==2.0.1
django-environ==0.4.5
pdftotext==2.1.5
django-crispy-forms==1.9.2
python-pptx==0.6.18
dj-database-url==0.5.0